+++ KnilUIKit/ListViewController/ListViewController.swift
@@ -39,7 +39,7 @@ public class ListViewController: UITableViewController {
         navigationItem.title = "AASA".localized()
         navigationItem.rightBarButtonItem = UIBarButtonItem(barButtonSystemItem: .add, target: self, action: #selector(presentAddingAASAAlertController))
 
-        NotificationCenter.default.addObserver(self, selector: #selector(appWillResignActive), name: .UIApplicationWillResignActive, object: nil)
+        NotificationCenter.default.addObserver(self, selector: #selector(appWillResignActive), name: UIApplication.willResignActiveNotification, object: nil)
 
         reloadData()
     }