+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [1.0.1(8)] - 2018-09-21
 ### Added
 - Support iOS 12 and Swift 4.2 #12